Greengrocerクラスの中でProductクラスのインスタンスを生成する
